requirements.txt
torch
torchvision
albumentations
pillow
numpy
pandas
matplotlib
tqdm
gymnasium
stable_baselines3
tensorboard
torcheval
torchmetrics
lpips
scipy